requirements.txt
Django
djangorestframework
django-bootstrap4==0.0.6
Pillow
psycopg2-binary

gunicorn
whitenoise
dj_database_url
Unipath